// src/client/types.ts
export type CompressionMethod = 'plain' | 'gzip' | (string & {});